fix(panasonic/camera/he_series): zoom should be a float

+++ drivers/panasonic/camera/he_series.cr
@@ -38,12 +38,12 @@ class Panasonic::Camera::HESeries < PlaceOS::Driver
 
   @invert : Bool = false
   @default_movement_speed : Int32 = 12
-  @presets = {} of String => NamedTuple(pan: Int32, tilt: Int32, zoom: Int32)
+  @presets = {} of String => NamedTuple(pan: Int32, tilt: Int32, zoom: Float64)
 
   def on_update
     @default_movement_speed = setting?(Int32, :default_movement_speed) || 12
     self[:inverted] = @invert = setting?(Bool, :invert_controls) || false
-    @presets = setting?(Hash(String, NamedTuple(pan: Int32, tilt: Int32, zoom: Int32)), :presets) || {} of String => NamedTuple(pan: Int32, tilt: Int32, zoom: Int32)
+    @presets = setting?(Hash(String, NamedTuple(pan: Int32, tilt: Int32, zoom: Float64)), :presets) || {} of String => NamedTuple(pan: Int32, tilt: Int32, zoom: Float64)
     self[:presets] = @presets.keys
   end
